Add fullcalendar
Reverted back to `date` because Hugo needs that to put the events in the
correct folders.

Using fullcalendar element. Fetches monthly events from the json file
and displays them. Links to the event page on our website.

We're pulling in the calendar in the head partial for now. Could be
added separately but keeping things simple for now.