remove console.log from head.html (#29)
There were two `console.log` calls in the project. One of them in the head.html file which I removed, and the other in the `build.js` file. I didn't remove the one in `build.js` as they're called only when running `yarn build`, and it makes sense to have a log in that function.